extract hbs render and rendering postprocessing to wrapper
