+++ src/renderer/html_handlebars/mod.rs
@@ -2,6 +2,7 @@
 
 pub use self::hbs_renderer::HtmlHandlebars;
 
+mod hbs_wrapper;
 mod hbs_renderer;
 mod helpers;
